// old/data/type.ts
export interface ITEM_STACK_DATA {
	/**
	 * TODO: 物品类型
	 */
	type: string;
	/**
	 * TODO: 物品数量
	 */
	amount?: number;
	/**
	 * TODO: 物品名称
	 */
	name?: string;
	/**
	 * TODO: 物品词缀
	 */
	lore?: string[];
	/**
	 * TODO: 物品动态属性
	 */
	property?: Map<string, boolean | number | string | server.Vector3>;
};

/**
 * TODO: 「奖杯」 - 信息存储
 */
export interface SCHEDULE_NOTE {
	/**
	 * TODO: 当前任务依赖的标签
	 */
	rely?: string[];
	/**
	 * TODO: 任务完成后添加的标签
	 */
	onDone?: string[];
	/**
	 * TODO: 任务完成后触发的引导
	 */
	prompt?: string;
	/**
	 * TODO: 任务的图标纹理
	 */
	texture?: string;
	/**
	 * TODO: 任务的文本描述
	 */
	refer: server.RawMessage[];
	/**
	 * TODO: 任务完成后的奖励
	 */
	reward?: ITEM_STACK_DATA[];
	/**
	 * TODO: 任务完成后的损耗
	 */
	attrition?: ITEM_STACK_DATA[];
	/**
	 * TODO: 任务完成后处理函数
	 */
	after?: (player: server.Player) => void;
	/**
	 * TODO: 任务完成前处理函数
	 */
	before?: (player: server.Player) => boolean;
};